Adds marketplace links to sidebar
Adds a "Marketplace" link to the main navigation and "My Listings" and "My Purchases" links to the account section of the sidebar.

Also adjusts the sidebar notification to use CSS classes instead of inline styles to control visibility based on the open state of the sidebar. This simplifies the logic and improves readability.