Conditionally renders lock controls
This commit ensures lock/unlock controls for files and folders are only rendered if the user is authorized. It also prevents lock state changes if the user isn't authorized.
This enhances the user experience by hiding features that users don't have permission to use and prevents errors.
Also, renames the "Latest marketplace listings" to "Latest templates listings" for accuracy.