Refactors auth logic and fixes race condition
Moves Supabase client initialization to the top level of the AuthSyncProvider and BuilderProvider components to ensure a single instance is used across the application. This prevents potential issues with multiple clients interfering with each other.

Also, prevents unnecessary chat data refreshes when the component is loading by adding a check for the loading state.